+++ internal/xerrors/issues.go
@@ -8,6 +8,8 @@ import (
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Issue"
 )
 
+const primaryKeyConstraintViolationCode = Ydb.StatusIds_StatusCode(2012)
+
 // NewWithIssues returns error which contains child issues
 func NewWithIssues(text string, issues ...error) error {
 	err := &errorWithIssues{
@@ -83,4 +85,17 @@ func iterate(
 		it(issue.GetMessage(), Ydb.StatusIds_StatusCode(issue.GetIssueCode()), issue.GetSeverity())
 		iterate(issue.GetIssues(), it)
 	}
+}
+
+func IsPrimaryKeyConstraintViolation(err error) (conflictDetected bool) {
+	var o *operationError
+	if !errors.As(err, &o) {
+		return false
+	}
+	iterate(o.Issues(), func(message string, code Ydb.StatusIds_StatusCode, severity uint32) {
+		if code == primaryKeyConstraintViolationCode {
+			conflictDetected = true
+		}
+	})
+	return conflictDetected
 }